updated labelFormatter to take the data point instead of the x, y values to give more context.
Added the ability to style the labels on bars in the BarChart.
The labelFormatter in BarChartConfig now provides the `ChartDataPoint` object instead of individual x and y values. This allows for more flexible label creation.

A new `labelTextStyle` property has been added to `BarChartConfig` to allow customization of the text style for bar labels.